requirements.txt
scikit-learn
pandas
matplotlib
llama-cpp-python
seaborn
numpy